Provide iex -S mix instead of mix iex
In order for iex to work reliably, we need it to pass special
parameters to the Erlang VM. Therefore, it cannot be started
via a regular Elixir script, as in `mix iex`.

Therefore, we need to use `iex -S mix` which first boots iex
and then run the given mix commands.

Closes #692