Rename the `args` variable to `attrs` or `opts` where appropriate
It is a good convention to have the name "args" refer to variables
that actually contain a list of arguments dispatched to a function
in some way.

When passing a keyword list to a record constructor,
it is referred to as "attributes" or "attrs".

When passing a list of arbitrary options,
it is referred to as "opts".