Set :proc_lib initial call on task
Sets the :proc_lib initial call to the {mod, fun, arity} of the task.

This allows :proc_lib.translate_initial_call(task.pid) to show the mfa
of a task process when debugging. Tools such as :observer and the erlang
shell make use of this function when getting information on a process.

This gives a useful initial call in the crash report sent to the error
logger when a task raises.