Add ability to slice a string given only a start index.
This mimics some languages (for instance C#), where you can ask for a
slice or substring given only a start index to slice from. This permits
easily trimming N many characters from the beginning of a string without
having to know or specify the length in order to make sure the remainder
of the string is returned.