Make IO.ANSI.Docs readable on white background
The `default_options` for `IO.ANSI.Docs` uses bright yellow for
documentation headers and titles.

This is very nice and readable when using a terminal with a dark
background. However, when using a terminal with a bright (e.g. white)
background, those parts are unfortunately illegible.

This patch removes the `:bright` part from the `default_options`, to
make the defaults better suited to more people.