Fix GenServer.cast/2 docs
Removes ambiguity about what occurs when an atom is used to reference a
GenServer.

Clarifies comment about another process being used to send a cast in
OTP. Another process is only used if there is no connection the node and
only casts to other nodes are not guaranteed to arrive in order.