Refine parsing of no parens calls
Reimplements 7f528aa1c035bf5f0e62fe20b6e3975c0084f395, which was
reverted due to having introduced a bug.  This commit avoids that bug
and includes a regression test for it.  Hat-tip to José Valim for
finding a way to refactor this PR so as to simplify it greatly.

Closes #3074