Refine parsing of no parens calls
Reimplements e859462df23d9c20bd47073bbcd042a592f5f98a, which was
reverted due to having introduced a bug.  This commit avoids that bug
and includes a regression test for it.  Hat-tip to José Valim for
finding a way to refactor this PR so as to simplify it greatly.

Closes #3074