Fix doctest :only, so it accepts multiple functions
There is a bug in `doctest` with the `:only` option, such that it only works if
the list is of size 1. If you pass two functions to the option, `filter_by_opts`
will always be empty, since the function is checked with `Enum.all?`.

`Enum.any?` returns false for an empty list. Thus the extra `Enum.empty?` check.