Add p/0 to IEx.Helpers
A lot of other language's interpreters (e.g. ruby, node, off the top of my head) allow you to use the result of the last evaluated expression via a shortcut such as `_`. This commit adds a helper `p/0` to `IEx.Helpers` which allows you to do the same in IEx. `_` would have been cool but of course that is not available.