No longer translate arguments to `@on_definition`
Breaking change: the arguments to the `@on_definition` hook are no
longer expanded and translated, but are passed in their original quoted
form.  The only exception is module attributes (e.g. `@my_constant`),
which are expanded.

The reason for this change is that this hook is used for generating
functions' documentation, and the previous behavior was causing
incorrect signatures in the documentation.  See #3143 for details.

Closes #3143