Refactor File! methods to avoid going through IO.chardata_to_string twice
All the non-! methods pass the pass argument through IO.chardata_to_string,
and we are doing this work twice for the ! methods since we convert once
before delegating to the non-! versions.

This moves the conversion to the exception case only, which should be
far less common, thus avoiding going through the duplicated call.