Use agents to store Mix config
Mix config was stored in a variable (`var!(config, Mix.Config)`) which was
initialized at every `use Mix.Config`. Using a variable worked well but didn't
allow pretty much anything outside "straight" code. For example, you could do
this:

    config :foo, bar: :baz

    if 3 + 2 == 5
      config :foo, [math: :check]
    end

but you could not do this:

    f = fn ->
      config :foo, bar: :baz
    end
    f.()

since functions can't leak their scope into the outside scope, so changes to the
`config` variable were not "persisted". `for` comprehensions were not allowed in
the same way.

This commit introduces a `Mix.Config.Agent` module so that configuration can be
stored inside an agent. An agent is started every time that `use Mix.Config` is
called, and is stopped by `Mix.Config.read!/1` (which is the basic unit for
reading configuration; everything else should use it).