Showing example of timeout for Task.await
It took me a bit of time to figure out how to handle a timeout. Figured I'd add an example here for future reference. 

I wasn't sure if I should include this near the module doc or in the function doc. I figured it would be best near the function. 

Let me know what you think!