Move terminate reason to second line in logs
With default settings logger could truncate the log before the
terminate reason if there is a large term beforehand, such as a large
message in a GenServer termination report. Therefore move terminate
reason (the most important information) as near the front of the
message as possible so that as little of it as possible is truncated.