Favor explicitness in Mix.Utils.read_path with rebar/archive installs
The previous code had a bunch of indirection and if/else
clauses because we tried to shoe horn all different concerns
from rebar, archives, reading and copying in a single place.

The new code has a simple Mix.Utils.read_path/2 which returns
the exact failure reasons to the calling code, allowing us to
provide a better easy experience regardless of the failure.