Only recompile compile time dependencies in mix projects
Prior to this patch, if module A invoked B, regardless of
where, A would recompile if B was changed.

Since the previous commit, we are now able to distinguish
between compile time dependencies (macros and calls outside
of a function) from runtime ones (everything else).

Now when traversing the dependency tree, runtime dependencies
since marks a given module as stale. If there is, however, a
compile time dependency on any stale module, the module that
depends on the stale dependency is recompiled.

To be more precise, imagine this flow:

    A [calls macro in] B [calls function in] C

Previously, A, B and C would recompile if C changed. Now only
A will.