Tweak the docs for the Inspect protocol
The docs for the Inspect protocol mentioned that if there's an error
while inspecting a struct (for which the Inspect protocol has been
implemented), then the struct is printed with a raw representation.

This commit makes it clear that an `ArgumentError` is raised and the
struct is printed raw in the error message.