Slight imporvement in doctest error line numbers
A bit hacky but better than previous IMO.

As of now the line numbers returned from `:beam_lib.chunks` are not
accurate.
For docs it refers to the first line of the function, for moduledocs it refers
to the first line of the module. The hack I've written here only works if the
docstring ends right before the function head and if the moduledoc is on the first
line.