fix(Kernel): check arguments on if
The following things are now checked when using `Kernel.if`:

 * There is only one `do` key
 * There is only one `else` key
 * Only `do` and `else` keys exist
 * One of `do` or `else` must exist

Any calls to the if macro that do not satisfy these conditions raise an
ArgumentError with a relevant error message.

Since if is used in Keyword, it is not possible to use
`Keyword.has_key?` as it depends on `if` - recursion is used instead.