fix(Kernel): ensure unless arguments are valid
The `unless` macro now has the same constraints as the `if` macro.
Without these checks, the `unless` macro would always fail based on the
fact that `if` requires a `do` block and `unless` calls `if` with the
arguments inverted.

To ensure the correct error message is displayed, a `type` string is
passed through to ensure the correct error message is displayed.