Make a distinction between touching and updating the lock manifest
Previously, deps.compile would update the parent project
lock manifest, which updated its Elixir version and made
it impossible to detect later on that we were upgrading
from a previous Elixir version.

This commit adds the ability to both touch and update
the manifest file and change deps.compile to only touch
it.