Allow size and map key variable matching in for and fn
We are treating map keys as patterns, so the pin
operator is required when matching on those.

We are treating bitstring specifiers as values,
so the pin operator is not required when matching
on those.

It is debatable if bitstring specifiers should
also be considered matches and require the pin
operator in such occasions, the reason why we have
decided to keep it as value is two fold:

1. Elixir 1.0 used them as values, so it is
   backwards compatible;

2. Supporting them as patterns introduces
   ambiguity since they are never matched,
   for example, we would need to write:

     x = 8
     <<a, b::size(^x)>> # valid
     <<x, y::size(x)>>  # valid
     <<a, b::size(x)>>  # invalid

   The last example contradicts regular matching
   where it would be valid with or without the pin
   operator. Notice this problem does not occur with
   maps because maps can never bind on a variable
   introduced in the pattern itself.