Support passing io_device to File.copy/3
Technically it is possible to pass an io_device to File.copy/3 as either
the source, destination, or both. However, all input was being converted
to a string, which fails for any non-binary data and prevents
`:file.copy` from being reached.

This updates the typespec and conditionally converts chardata to a
string. The file error message now inspects both source and destination
for uniform logging.