Make GenServer.cast/2 always return :ok. Closes #4251
This unifies the behavior of `GenServer.cast/2` to always
return :ok and never raise an `ArgumentError`, regardless
of how it is called.

Previously an `ArgumentError` would be be raised only when
called with a nonexistant local process name. In all other cases
which are calling using `:global`, `:via` or `{name, node}`
it would previously always return `:ok`.

The new behavior is now constistant with `:gen_server.cast/2`.

This reverts the change partially introduced in 201450bcd726659f3055cad074031fdeee2899b0.

See #4251 for the full discussion.