Suppress unused var warnings when removing Logger macros from the AST
Right now, code like this:

    defmodule MyMod do
      require Logger
      def my_fun(arg) do
        Logger.debug(arg)
        :ok
      end
    end

will generate an unused variable warning for the `arg` variable in
`my_fun/1` if the log level is higher than :debug. This happens because
in such cases, we completely remove the AST of the Logger call, so the
`arg` variable is effectively never used. This may seem magical to the
user.

With this commit, a Logger call which would have been removed from the
AST before now generates a series of `_ = x` expressions, one for each
variable passed to the Logger macro.