Error out when passing :__aliases__ to Macro.pipe/3
Right now, this funny thing happens:

    iex> Macro |> Env
    Macro.Env

This happens because `|>/2` calls `Macro.pipe/3` under the hood, and
`Macro.pipe/3` inserts the left-hand side at the beginning of the list
of arguments in the AST for the function call on the right-hand
side. Aliases, however, look exactly like a call to the __aliases__
function:

    iex> quote do: Macro.Env
    {:__aliases__, [alias: false], [:Macro, :Env]}

and that's why `Macro |> Env` becomes `Macro.Env`. With this commit, we
explicitly take care of `:__aliases__` in a `Macro.pipe/3` call, so that
we can error out in such cases:

    iex> Macro |> Env
    ** (ArgumentError) cannot pipe Macro into Env, ...