Handle non-called anonymous functions in Macro.pipe/3
Right now, this nasty error pops out:

    iex> :foo |> (fn x -> x end)
    ** (CompileError) iex:1: expected -> clauses in fn

This happens because of what we do in `Macro.pipe/3`, which is: we take
the AST of a function call as the second argument and insert the first
argument in the list of arguments for the function call, at the position
specified by the third argument. The AST for a `fn` looks exactly like
the AST of a function call (to the `:fn` function):

    iex> quote do: fn :ok -> :ok; :err -> :err end
    {:fn, [], [{:->, [], [[:ok], :ok]}, {:->, [], [[:error], :error]}]}

It's easy to see that the "arguments" to `:fn` are the `->` clauses in
the function. This means that doing `:foo |> (fn x -> x end)` basically
adds another clause to the function:

    iex> Macro.pipe(:foo, quote(do: fn :ok -> :ok; :err -> :err end), 0)
    {:fn, [], [:foo, {:->, [], [[:ok], :ok]}, {:->, [], [[:err], :err]}]}

and this makes `:elixir_fn.expand/3` blow up because it doesn't know how
to handle the `:foo` clause (because it's not a clause!).

With this commit, we blow up early in `Macro.pipe/3` when we see a
`{:fn, _, _}` call:

    iex> :foo |> (fn x -> x end)
    ** (ArgumentError) cannot pipe :foo into an anonymous function
    without calling the function; use :foo |> (fn x -> x end).()