Fix and improve Enum.sum/1 for ranges
Lastest changes introduced in Enum.sum changed the type of the return value
(from integer to float) when argument was a range. This commit fixes this.

Adds a spec when dealing with Range.t

It simplifies the function by making it more recursive,
and also deals with ranges where first and last are equal

Tests added