Remove the use of the word `pad` from chunk
We will now take the word `pad` to mean something used for
completion and cycled until the number of desired elements
are given.

Given this is not the behaviour in `chunk`, we renamed the
argument from `pad` to `leftover`.

This patch does not change the behaviour, it just amends
the documentation to use proper names.