Friendly error message for duplicate ExUnit test names (#4613)
Previously, duplicate test names would output a warning like:

  warning: this clause cannot match because a previous clause at line 6 always matches

In order to understand the warning, the user had know that ExUnit defines
functions based on the test names. This was an implementation detail that the
user shouldn't have to know about. Since the second test definition is useless,
make the definition of it fail immediately without running the tests and
explicitly tell the user why. The new error looks like:

  ** (ExUnit.DuplicateTestError) A test named "foo" is already defined in Elixir.DupTest.
      test/dup_test.exs:6: (module)
      (stdlib) erl_eval.erl:670: :erl_eval.do_apply/6