Disable overriding the async tag
The `async` config is meant to be set only by the time `ExUnit.Case`
is used. With these changes we make sure a proper error will be
raised if we eventually try to override it, either at the setup step,
for the whole case as a moduletag, or as a regular test tag.