Use square brackets for IPv6 in URI.to_string/1
According to the grammar at
https://tools.ietf.org/html/rfc3986#appendix-A, a host can have colons
in it only if it's a "IPv6" or a "IPvFuture"; in both cases, it's
expected to be surrounded by square brackets. This means that when
converting a URI to a string in URI.to_string/1, we can simply surround
a host with square brackets if it contains one or more colons.