Keep brackets in the authority of URIs for IPv6 hosts
Right now, when parsing a URI, we strip square brackets out of
hosts (meaning they're IPv6 hosts) and then use the stripped host to
rebuild the authority that will end up in the %URI{} struct. This means
that, for example:

    URI.parse("foo://[::]:1234").authority == ":::1234"

which is not good. In this commit, I fixed this by not rebuilding the
authority using the new host in URI.parse/1. The authority will now
include brackets around the host, but the :host field in the %URI{}
struct is still stripped.