Clarify the use of `async: true` (#4692)
The documentation here was a bit confusingly worded to make it sound
like with `async: true` each `test` declaration in the case would run
asynchronously. I've tried to reword it and make it clear that the flag
is set for the test case itself.

One example of the confusion is issue #3580. I found that thread when
trying to find the same answer for why six tests with
`:timer.sleep(1000)` were running in serial within a single test case.