Fix failing tests for System.cmd/3 on windows
As mentioned in https://github.com/elixir-lang/elixir/issues/4753 Windows echo is not a separate executable
it is a builtin to the shell (cmd.exe).

Create separate System.cmd/3 tests for windows which use the cmd.exe the echo a message.
Also change after block of test "cmd/2 with absolute and relative paths win" - sometimes File.rm_rf!
failed with 'permission denied'.