Use heuristic for compiler errors (#4778)
Previously when all files being compiled get stuck waiting on
something else, we would release those files in random order.

This commit changes this to release files in groups based on
the module they are waiting on. We pick the module being
depended on with less edges, as it is mostly likely being the
cause of conflict. This may not always be true though: for example,
if there is a macro injecting code into multiple modules and such
code becomes faulty, now multiple modules are waiting on
this same module.

However, we have to pick one or the other, and we
decided to pick the one with fewer dependencies.