Raise an error when trying to make private macros overridable
This is not supported as of today because of bugs regarding defoverride
and macros, but we decided to not support overriding private macros at
all so let's raise a nice error message.

For the record, supporting overridable private macros would be very
annoying since we rely on function names and the such for implementing
overridable stuff, and for private macros we just generate functions on
the fly.