Stop appending a trailing dot when autocompleting modules in IEx.
I find this behavior to be surprising and annoying. I regularly
type `h [first few chars of mod name]<tab><enter>` in IEx to read
the read the moduledoc of a particular module, and the trailing
dot gets in the way every time.

Instead, we append a dot when you hit tab after the module name is
already complete--so if you hit tab twice, it will autocomplete the
module name and then append a trailing dot.