Generate error for no pid given to Genserver.reply
Prior to this change it is possible to send any two-pair tuple to
`GenServer.reply/2` and it will happily reply with `:ok`.  This
leads to some hard to track-down bugs if you pass it the wrong
data.  With this small check developers may be able to get a bit
more of a heads up to what is going wrong sooner.

For more information on how I came accross this:
https://github.com/benfalk/data_pool/commit/dcf81d1a87a363952715deeed3caa18ec0efccd9