Reword some docs in String because of an Earmark bug
Currently Earmark parses both `\` as a monospaced font backslash but
considers the following backtick as escaped by the backslash and thus
messes up the following Markdown and warns about a missing closing
backtick. Similar for `\\`, where we have two backslashes in monospaced
font but the same behaviour regarding backticks. For now, the easiest
solution is to just replace the backslash with the word "backslash"
until the bug is fixed in Earmark (see
https://github.com/pragdave/earmark/issues/78).