Add Diameter compiler for mix
This commit brings a simple support of Diameter compiler.
I was inspired by `rebar_dia_compiler` and it should work the same way
expect `dia_first_files` option.

There are some unresolved questions:
1) Should I add the `--force` option like the others compilers have.
2) Diameter sources have the inherits and it requires the correct file
order.  If `b` inherits `a` then `a` should be compiled before `b`. In
`rebar_dia_compiler` it is solved by `dia_first_files` but it is not
possible to add that kind of option to this compile so I may suggest to
use alphabetic order to naming dia sources for now.
3) Diameter compiler generates `erl` and `hrl` file. It is possible in
Elixir to work with records from `hrl` file but I don't know a good way to
work with defined constants.